+++ mail_check/CHANGELOG.md
@@ -4,4 +4,7 @@
 
 ## [0.0.3] - 9 Aug 2023
 * Enhance suggestion email
-* Support check email validation with regex 
+* Support check email validation with regex
+
+## [0.0.4] - 9 Aug 2023
+* Support warning reason (exist special characters or not found domain)